# Plugins/BTCPayServer.Plugins.SatoshiTickets/Resources/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SatoshiTickets Greenfield API",
    "description": "REST API for the SatoshiTickets BTCPay Server plugin. Provides endpoints for managing events, ticket types, tickets, orders, and check-ins.\n\nAll endpoints are scoped to a specific store and require a Greenfield API key with appropriate permissions.\n\n**Authentication:** Pass your API key in the `Authorization` header:\n```\nAuthorization: token YOUR_API_KEY\n```\n\nCreate an API key under **Account > Manage Account > API Keys** in BTCPay Server.",
    "version": "1.2.2",
    "contact": {
      "name": "SatoshiTickets",
      "url": "https://github.com/TChukwuleta/BTCPayServerPlugins"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "/api/v1/stores/{storeId}/satoshi-tickets",
      "description": "Current BTCPay Server instance",
      "variables": {
        "storeId": {
          "default": "your-store-id",
          "description": "The store ID"
        }
      }
    }
  ],
...
  "paths": {
...
    "/events/{eventId}/tickets": {
      "get": {
        "tags": ["Tickets"],
        "summary": "List tickets",
        "description": "Returns settled (paid) tickets for an event. Supports text search across ticket number, name, and email.",
        "operationId": "getTickets",
        "parameters": [
          {
            "$ref": "#/components/parameters/eventId"
          },
          {
            "name": "searchText",
            "in": "query",
            "description": "Search in `txnNumber`, `firstName`, `lastName`, `email`, `ticketNumber`.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of tickets",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TicketData"
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/EventNotFound"
          }
        },
        "security": [{ "API_Key": ["btcpay.store.canmodifystoresettings"] }]
      }
    },
...
    "/events/{eventId}/tickets/{ticketNumber}/check-in": {
      "post": {
        "tags": ["Tickets"],
        "summary": "Check-in ticket",
        "description": "Marks a ticket as checked in. Accepts either the full `ticketNumber` (e.g. `EVT-abc123-260615-xK9mP2qR`) or just the short `txnNumber` (e.g. `xK9mP2qR`).",
        "operationId": "checkinTicket",
        "parameters": [
          {
            "$ref": "#/components/parameters/eventId"
          },
          {
            "name": "ticketNumber",
            "in": "path",
            "description": "Full ticket number or short transaction number.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Check-in result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckinResultData"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/EventNotFound"
          },
          "422": {
            "description": "Check-in failed (invalid ticket or already checked in)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GreenfieldError"
                }
              }
            }
          }
        },
        "security": [{ "API_Key": ["btcpay.store.canmodifystoresettings"] }]
      }
    },
    "/events/{eventId}/purchase": {
      "post": {
        "tags": ["Orders"],
        "summary": "Create ticket purchase",
        "description": "Creates a ticket order programmatically and returns the BTCPay invoice checkout URL. Use this to integrate ticket sales with external systems (e.g. WooCommerce).\n\n**Flow:** Validates event, ticket types, and availability → creates Order and Tickets in DB → creates BTCPay invoice → returns checkout URL for payment.",
        "operationId": "createPurchase",
        "parameters": [
          {
            "$ref": "#/components/parameters/eventId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePurchaseRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Purchase created — returns order ID, invoice ID, and checkout URL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurchaseResponse"
                }
              }
            }
          },
          "404": {
            "description": "Event or ticket type not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GreenfieldError"
                }
              }
            }
          },
          "422": {
            "description": "Validation error (insufficient quantity, recipients count mismatch, event not active, invalid email, etc.)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GreenfieldError"
                }
              }
            }
          }
        },
        "security": [{ "API_Key": ["btcpay.store.canmodifystoresettings"] }]
      }
    },
    "/events/{eventId}/create-tickets-offline": {
      "post": {
        "tags": ["Orders"],
        "summary": "Create tickets (offline payment)",
        "description": "Creates tickets without a BTCPay invoice. Use when payment was already received outside BTCPay (e.g. bank transfer, cash, COD). Validates event and ticket types, creates Order and Tickets with `PaymentStatus=Settled`, updates stock (QuantitySold), and sends ticket emails to recipients.\n\n**Difference from purchase:** Does not create an invoice or checkout URL; the order is immediately settled.",
        "operationId": "createTicketsOffline",
        "parameters": [
          {
            "$ref": "#/components/parameters/eventId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTicketsOfflineRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tickets created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTicketsOfflineResponse"
                }
              }
            }
          },
          "404": {
            "description": "Event or ticket type not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GreenfieldError"
                }
              }
            }
          },
          "422": {
            "description": "Validation error (insufficient quantity, recipients count mismatch, event not active, invalid email, etc.)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GreenfieldError"
                }
              }
            }
          }
        },
        "security": [{ "API_Key": ["btcpay.store.canmodifystoresettings"] }]
      }
    },
    "/events/{eventId}/orders": {
      "get": {
        "tags": ["Orders"],
        "summary": "List orders",
        "description": "Returns settled (paid) orders for an event, each with nested tickets.",
        "operationId": "getOrders",
        "parameters": [
          {
            "$ref": "#/components/parameters/eventId"
          },
          {
            "name": "searchText",
            "in": "query",
            "description": "Search in `invoiceId`, ticket `txnNumber`, `firstName`, `lastName`, `email`.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of orders with tickets",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrderData"
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/EventNotFound"
          }
        },
        "security": [{ "API_Key": ["btcpay.store.canmodifystoresettings"] }]
      }
    },
    "/events/{eventId}/orders/{orderId}/tickets/{ticketId}/send-reminder": {
      "post": {
        "tags": ["Orders"],
        "summary": "Send ticket reminder email",
        "description": "Re-sends the ticket registration email to the ticket holder. The `ticketId` is the ticket's UUID (`id` field), not the `ticketNumber`.",
        "operationId": "sendReminder",
        "parameters": [
          {
            "$ref": "#/components/parameters/eventId"
          },
          {
            "name": "orderId",
            "in": "path",
            "description": "The order ID.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ticketId",
            "in": "path",
            "description": "The ticket UUID (`id` field).",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Email sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Ticket details have been sent to the recipient via email"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Order or ticket not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GreenfieldError"
                }
              }
            }
          },
          "422": {
            "description": "Email SMTP not configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GreenfieldError"
                },
                "example": {
                  "code": "email-not-configured",
                  "message": "Email SMTP settings are not configured. Configure email settings in the store admin."
                }
              }
            }
          },
          "500": {
            "description": "Email sending failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GreenfieldError"
                },
                "example": {
                  "code": "email-send-failed",
                  "message": "Failed to send ticket email to: user@example.com"
                }
              }
            }
          }
        },
        "security": [{ "API_Key": ["btcpay.store.canmodifystoresettings"] }]
      }
    }
  },
  "components": {
    "securitySchemes": {
      "API_Key": {
        "type": "apiKey",
        "description": "BTCPay Server Greenfield API Key. Pass as `Authorization: token <key>`. Create under Account > Manage Account > API Keys. Required permission: `btcpay.store.canmodifystoresettings`.",
        "name": "Authorization",
        "in": "header"
      }
    },
    "parameters": {
      "eventId": {
        "name": "eventId",
        "in": "path",
        "description": "The event ID.",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
...
    },
    "responses": {
      "EventNotFound": {
        "description": "Event not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/GreenfieldError"
            },
            "example": {
              "code": "event-not-found",
              "message": "The event was not found"
            }
          }
        }
      }
    },
    "schemas": {
...
      "CheckinResultData": {
        "type": "object",
        "description": "Result of a check-in attempt.",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Whether the check-in succeeded."
          },
          "errorMessage": {
            "type": "string",
            "nullable": true,
            "description": "Error message if check-in failed (e.g. already checked in)."
          },
          "ticket": {
            "allOf": [
              { "$ref": "#/components/schemas/TicketData" }
            ],
            "nullable": true,
            "description": "The ticket data, or `null` if the ticket was invalid."
          }
        }
      },
...
      "GreenfieldError": {
        "type": "object",
        "description": "Standard Greenfield API error.",
        "properties": {
          "code": {
            "type": "string",
            "description": "Machine-readable error code."
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message."
          }
        }
      },
...
    }
  }
}